Fix references to parquet; restructure
