+++ .gitignore
@@ -6,3 +6,4 @@ ex_darshan_logs
 XPDaMoDa
 parquet/
 parquet2/
+__pycache__
