+++ src/wfmeta_darshan/objs/__init__.py
@@ -1,0 +1,1 @@
+from . import colls
